chore(dir): #1082 rename lib/ to addon/
