+++ .eslintignore
@@ -1,8 +1,8 @@
 data/content/
 logs/
 firefox/
-lib/vendor.bundle.js
+addon/vendor.bundle.js
 activity-streams-env/
 !.*
-lib/SimpleStorage.js
+addon/SimpleStorage.js
 test/test-SimpleStorage.js
