+++ addon/ActivityStreams.js
@@ -13,21 +13,21 @@ const privateBrowsing = require("sdk/private-browsing");
 const windows = require("sdk/windows").browserWindows;
 const prefService = require("sdk/preferences/service");
 const ss = require("sdk/simple-storage");
-const {Memoizer} = require("lib/Memoizer");
-const {PlacesProvider} = require("lib/PlacesProvider");
-const {SearchProvider} = require("lib/SearchProvider");
-const {ShareProvider} = require("lib/ShareProvider");
-const {TabTracker} = require("lib/TabTracker");
-const {PreviewProvider} = require("lib/PreviewProvider");
-const {RecommendationProvider} = require("lib/RecommendationProvider");
-const {TelemetrySender} = require("lib/TelemetrySender");
-const {PerfMeter} = require("lib/PerfMeter");
-const {AppURLHider} = require("lib/AppURLHider");
+const {Memoizer} = require("addon/Memoizer");
+const {PlacesProvider} = require("addon/PlacesProvider");
+const {SearchProvider} = require("addon/SearchProvider");
+const {ShareProvider} = require("addon/ShareProvider");
+const {TabTracker} = require("addon/TabTracker");
+const {PreviewProvider} = require("addon/PreviewProvider");
+const {RecommendationProvider} = require("addon/RecommendationProvider");
+const {TelemetrySender} = require("addon/TelemetrySender");
+const {PerfMeter} = require("addon/PerfMeter");
+const {AppURLHider} = require("addon/AppURLHider");
 const am = require("common/action-manager");
 const {CONTENT_TO_ADDON, ADDON_TO_CONTENT} = require("common/event-constants");
-const {ExperimentProvider} = require("lib/ExperimentProvider");
+const {ExperimentProvider} = require("addon/ExperimentProvider");
 const {Recommender} = require("common/recommender/Recommender");
-const {PrefsProvider} = require("lib/PrefsProvider");
+const {PrefsProvider} = require("addon/PrefsProvider");
 
 Cu.import("resource://gre/modules/XPCOMUtils.jsm");
 Cu.import("resource:///modules/NewTabURL.jsm");
