+++ addon/AppURLHider.js
