+++ addon/ColorAnalyzerProvider.js
