+++ addon/MetadataStore.js
