+++ addon/PerfMeter.js
