+++ addon/PlacesProvider.js
