+++ addon/PrefsProvider.js
