+++ addon/RecommendationProvider.js
