+++ addon/SearchProvider.js
