+++ addon/ShareManifests.js
