+++ addon/SimpleStorage.js
