+++ addon/TelemetrySender.js
