+++ addon/vendor-src.js
@@ -1,11 +1,11 @@
 // Note:
 // DO NOT import this file directly, as it needs to be processed by webpack.
 // Instead, import the bundle like this:
-// require("lib/vendor.bundle");
+// require("addon/vendor.bundle");
 try {
   platform_exports; // eslint-disable-line no-unused-expressions
 } catch (e) {
-  throw new Error("You are trying to import the wrong file. Import lib/vendor.bundle.js instead of lib/vendor-src.js");
+  throw new Error("You are trying to import the wrong file. Import addon/vendor.bundle.js instead of addon/vendor-src.js");
 }
 
 platform_exports = Object.assign(platform_exports, {
