+++ test/test-PreviewProvider.js
@@ -9,7 +9,7 @@ const {Loader} = require("sdk/test/loader");
 const loader = Loader(module);
 const httpd = loader.require("./lib/httpd");
 const {Cu} = require("chrome");
-const {PreviewProvider} = require("lib/PreviewProvider");
+const {PreviewProvider} = require("addon/PreviewProvider");
 const ALLOWED_PROTOCOLS = new Set(["http:", "https:"]);
 const DISALLOWED_HOSTS = new Set(["localhost", "127.0.0.1", "0.0.0.0"]);
 const URL_FILTERS = [
